requirements.txt
ezdxf
numpy
Pillow
tk
scipy
scikit-image
tk